\name{atw}
\alias{atw}
\docType{data}
\title{
Attitudes Toward Women Scale (ATW)
}
\description{
Responses by 201 undergraduates at the University of Illinois at Urbana-Champaign to Spence, Helmreich, and Stapp's (1974) 15-item Attitudes Toward Women Scale.
}
\usage{data(atw)}
\format{
  A data frame with 201 observations on the following 16 variables.
  \describe{
    \item{\code{subj}}{Subject identification number.}
    \item{\code{atw01}}{A woman should be as free as a man to propose marriage.}
    \item{\code{atw02}}{Women earning as much as their dates should bear equally the expense when they go out together.}
    \item{\code{atw03}}{Women should worry less about their rights and more about becoming good wives and mothers.}
    \item{\code{atw04}}{Women should assume their rightful place in business and all the professions along with men.}
    \item{\code{atw05}}{The intellectual leadership of a community should be largely in the hands of men.}
    \item{\code{atw06}}{Under modern economic conditions with women active outside the home, men should share in household tasks such as washing dishes and doing laundry.}
    \item{\code{atw07}}{Sons in a family should be given more encouragement to go to college than daughters.}
    \item{\code{atw08}}{It is insulting to women to have the "obey" clause still in the marriage service.}
    \item{\code{atw09}}{Swearing and obscenity are more repulsive in the speech of a woman than a man.}
    \item{\code{atw10}}{In general, the father should have greater authority than the mother in the bringing up of children.}
    \item{\code{atw11}}{Economic and social freedom is worth far more to women than acceptance of the ideal of femininity, which has been set up by men.}
    \item{\code{atw12}}{There are many jobs in which men should be given preference over women in being hired or promoted.}
    \item{\code{atw13}}{A woman should not expect to go to exactly the same places or to have quite the same freedom of action as a man.}
    \item{\code{atw14}}{Women should be given equal opportunity with men for apprenticeship in the various trades.}
    \item{\code{atw15}}{It is ridiculous for a woman to run a locomotive and for a man to darn socks.}
  }
}
\details{
Administered in the fall of 1999. Participants responded to each item using a 1-7 Likert scale in which 7 indicated \dQuote{Strongly Agree} and 1 indicated \dQuote{Strongly Disagree}.
}

\references{
Spence, J. T., Helmreich, R., & Stapp, J. (1975). Ratings of self and peers on sex role attributes and their relation to self-esteem and conceptions of masculinity and femininity. Journal of Personality and Social Psychology, 32(1), 29-39.
}
\examples{
data(atw)
}
\keyword{datasets}
